Add session handling

+++ src/Environment/Session/Controller.php
@@ -3,8 +3,8 @@
 namespace Environment\Session;
 
 
+use Latch\SessionManager;
 use Slim\Http\Response;
-use Environment\IController;
 
 /**
  * city-call
@@ -16,69 +16,33 @@ class Controller extends \Environment\Controller
     public function process(): Response
     {
         $request = $this->getRequest();
-        $arguments = $this->getArguments();
-        $reception = new  Reception($request, $arguments);
+        $isDelete = $request->isDelete();
 
-        $method = $this->getMethod();
         $response = $this->getResponse();
-        switch ($method) {
-            case self::DELETE:
-                $response = $this->delete($reception);
-                break;
-            case self::GET:
-                $response = $this->read($reception);
-                break;
-            case self::POST:
-                $response = $this->create($reception);
-                break;
-            case self::PUT:
-                $response = $this->update($reception);
-                break;
+        if ($isDelete) {
+
+            $arguments = $this->getArguments();
+            $reception = new  Reception($request, $arguments);
+
+            $response = $this->delete($reception);
         }
 
         return $response;
     }
 
+    /**
+     * @param Reception $reception
+     * @return Response
+     * @throws \Exception
+     */
     private function delete(Reception $reception): Response
     {
+        /** @var \Latch\ISession $item */
         $item = $reception->toDelete();
 
-        $logicResult = (new Logic($item, $this->dataPath))->delete();
+        $content = (new SessionManager($item, $this->getDataPath()))->finish();
 
-        $response = (new Presentation($this->response, $logicResult))->fromDelete();
-
-        return $response;
-    }
-
-    private function read(Reception $reception): Response
-    {
-        $item = $reception->toRead();
-
-        $logicResult = (new Logic($item, $this->dataPath))->read();
-
-        $response = (new Presentation($this->response, $logicResult))->fromRead();
-
-        return $response;
-    }
-
-    private function create(Reception $reception): Response
-    {
-        $item = $reception->toCreate();
-
-        $logicResult = (new Logic($item, $this->dataPath))->create();
-
-        $response = (new Presentation($this->response, $logicResult))->fromCreate();
-
-        return $response;
-    }
-
-    private function update(Reception $reception): Response
-    {
-        $item = $reception->toUpdate();
-
-        $logicResult = (new Logic($item, $this->dataPath))->update();
-
-        $response = (new Presentation($this->response, $logicResult))->fromUpdate();
+        $response = (new Presentation($this->getRequest(), $this->getResponse(), $content))->process();
 
         return $response;
     }
